+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - module-rails: Add itbash command with menu iterative to rails module
 - module-pnm: Remove phpmyadmin container from php-nginx-mysql module
 - core: Add after and before triggers for up command
+- module-pnm: Add Symfony framework suport for php-nginx-mysql module
 # Version 1.3.3
 - module-rails: Fix command import_dump
 - module-rails: Fix command dump